mmironyuk-lesson-7 move apple creating in own class
