+++ mironyuk/7/src/org/serieznyi/shop/fabric/AppleGarden.java
@@ -1,5 +1,6 @@
 package org.serieznyi.shop.fabric;
 
+import org.serieznyi.shop.factory.AppleFactory;
 import org.serieznyi.shop.item.food.Apple;
 
 import java.util.ArrayList;
@@ -10,8 +11,13 @@ import java.util.concurrent.ThreadLocalRandom;
 import static java.lang.String.format;
 
 public class AppleGarden {
-    private int counter = 1;
+    private final AppleFactory appleFactory;
     private final ThreadLocalRandom random = ThreadLocalRandom.current();
+
+    public AppleGarden(AppleFactory appleFactory)
+    {
+        this.appleFactory = appleFactory;
+    }
 
     public void fillShopWithApples(Collection<? super Apple> shop)
     {
@@ -29,27 +35,9 @@ public class AppleGarden {
         Collection<Apple> apples = new ArrayList<>();
 
         for (int i = 0 ; i < random.nextInt(2, 10) ; i++) {
-            apples.add(makeApple());
+            apples.add(this.appleFactory.create());
         }
 
         return apples;
-    }
-
-    private Apple makeApple()
-    {
-        return new Apple(
-                "Яблоко №" + counter++,
-                random.nextInt(10, 50),
-                random.nextInt(10, 40),
-                random.nextInt(5, 10),
-                nextAppleColor()
-        );
-    }
-
-    private Apple.Color nextAppleColor()
-    {
-        Apple.Color[] colors = Apple.Color.values();
-
-        return colors[random.nextInt(0, colors.length)];
     }
 }
