+++ mironyuk/7/src/org/serieznyi/shop/factory/ItemFactory.java
@@ -1,0 +1,7 @@
+package org.serieznyi.shop.factory;
+
+import org.serieznyi.shop.item.ShopItem;
+
+public interface ItemFactory {
+    public ShopItem create();
+}
